+++ src/controllers/web/server/Oauth2AccessTokenAction.php
@@ -4,7 +4,6 @@ namespace rhertogh\Yii2Oauth2Server\controllers\web\server;
 
 use rhertogh\Yii2Oauth2Server\controllers\web\Oauth2ServerController;
 use rhertogh\Yii2Oauth2Server\controllers\web\server\base\Oauth2BaseServerAction;
-use rhertogh\Yii2Oauth2Server\exceptions\Oauth2ServerHttpException;
 use rhertogh\Yii2Oauth2Server\helpers\Psr7Helper;
 use Yii;
 use yii\web\HttpException;