+++ src/interfaces/controllers/web/Oauth2OidcControllerInterface.php
@@ -15,4 +15,10 @@ interface Oauth2OidcControllerInterface
      * @since 1.0.0
      */
     public const ACTION_NAME_USERINFO = 'userinfo';
+
+    /**
+     * Name for the Oauth2OidcEndSessionAction in the controller's actions
+     * @since 1.0.0
+     */
+    public const ACTION_END_SESSION = 'end-session';
 }